Add missing ranks for Actinopterygii
Also a quick note that some of the groupings used for Actinopterygii weren't the dominant groups on Wikipidia and were mostly chosen because of the branching factor (less then seven children). An original grouping used Percomorpharia, Carangimopharia, Gobiomopharia & Syngnathomorpharia but these were adjusted slightly to better match what was in Wikipedia. Some of these are discussed in this 2016 paper "Phylogenetic placement of enigmatic percomorph families (Teleostei: Percomorphaceae)":

- https://www.sciencedirect.com/science/article/abs/pii/S1055790315003097
- https://currents.plos.org/treeoflife/article/the-tree-of-life-and-a-new-classification-of-bony-fishes/