Make search input for iphone
Previously there was a display issue for the search input in safari so it was marked as a text input using a safari only CSS rule. This removed some usability from the iphone too. It seems like CSS styling has now updated the input in safari so this is CSS hack is no longer needed.